requirements.txt
pyserial==3.5
crcmod==1.7
